requirements.txt
beautifulsoup4==4.6.3
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
google==2.0.1
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-language==1.0.2
googleapis-common-protos==1.5.3
grpcio==1.15.0
gsearch==1.6.0
idna==2.7
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
jsonmerge==1.5.1
jsonschema==2.6.0
lxml==4.2.5
MarkupSafe==1.0
newspaper3k==0.2.8
nltk==3.3
Pillow==5.2.0
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
rsa==4.0
six==1.11.0
tinysegmenter==0.3
tldextract==2.2.0
urllib3==1.23
Werkzeug==0.14.1
